+++ packages/fighting-design/page-header/src/page-header.vue
@@ -1,6 +1,8 @@
 <script lang="ts" setup name="FPageHeader">
   import { computed } from 'vue'
-  import { FIcon } from '../../icon'
+  import { sizeChange } from '../../_utils'
+  import fIconArrowLeftVue from '../../_components/svg/f-icon-arrow-left.vue'
+  import { FSvgIcon } from '../../svg-icon'
   import { Props, Emits } from './page-header'
   import type { ComputedRef } from 'vue'
   import type {
@@ -17,15 +19,15 @@
 
   const rightClassList: ComputedRef<b> = computed((): b => {
     return [
-      'f-page-header-main',
-      { 'f-page-header-main-center': prop.titleCenter }
+      'f-page-header__main',
+      { 'f-page-header__main-center': prop.titleCenter }
     ] as const
   })
 
   const rightTitleClassList: ComputedRef<b> = computed((): b => {
     return [
-      'f-page-header-main-title',
-      { 'f-page-header-main-title-bold': prop.titleBold }
+      'f-page-header__main-title',
+      { 'f-page-header__main-title-bold': prop.titleBold }
     ] as const
   })
 </script>
@@ -33,13 +35,13 @@
 <template>
   <header class="f-page-header">
     <!-- 左侧部分 -->
-    <div class="f-page-header-left" @click="handleClick">
-      <div v-if="icon || $slots.icon" class="f-page-header-icon">
+    <div class="f-page-header__left" @click="handleClick">
+      <f-svg-icon :size="sizeChange(iconSize)">
         <slot name="icon">
-          <f-icon v-if="icon" :icon="icon" :size="iconSize" />
+          <component :is="icon || fIconArrowLeftVue" />
         </slot>
-      </div>
-      <div class="f-page-header-text">{{ backText }}</div>
+      </f-svg-icon>
+      <div class="f-page-header__text">{{ backText }}</div>
     </div>
 
     <!-- 右侧部分 -->
@@ -52,7 +54,7 @@
       <!-- 副标题 -->
       <div
         v-if="subtitle || $slots.subtitle"
-        class="f-page-header-main-subtitle"
+        class="f-page-header__main-subtitle"
       >
         <slot name="subtitle">{{ subtitle }}</slot>
       </div>